Add files via upload
The app was built with Pyhton, Javascript (Vanilla), and HTML5/CSS. There are more efficient and better ways to do this, such as using Python libraries and frameworks like Django and Flask, but I chose the basic approach. I know the fundamentals of all three languages, but Python is my preferred language, and I use Jupyter notebook for data extraction and manipulation.